added format package, removed from main.go
