+++ main.go
@@ -1,7 +1,8 @@
-package main
+package elasticsearch_data_loader
 
 import (
 	"bytes"
+	"elasticsearch-data-loader/format"
 	"encoding/json"
 	"flag"
 	"fmt"
@@ -9,9 +10,7 @@ import (
 	"log"
 	"net/http"
 	"os"
-	"strings"
 
-	"regexp"
 	"sync"
 	"time"
 
@@ -26,10 +25,6 @@ var (
 	alphakeyURL = "http://chs-alphakey-pp.internal.ch"
 	esDestURL   = "http://localhost:9200"
 	mongoURL    = "mongodb://envb.mongo.ch.gov.uk:15107"
-)
-
-var (
-	nonWordEndRegex = regexp.MustCompile(`[^A-Za-z0-9_]+$`)
 )
 
 var (
@@ -51,80 +46,6 @@ var (
 	skipChannel   = make(chan int)
 	semaphore     = make(chan int, 5)
 )
-
-var companyNameEndings = [...]string{
-	"AEIE",
-	"ANGHYFYNGEDIG",
-	"C.B.C",
-	"C.C.C",
-	"C.I.C",
-	"CBC",
-	"CBCN",
-	"CBP",
-	"CCC",
-	"CCG CYF",
-	"CCG CYFYNGEDIG",
-	"CIC",
-	"COMMUNITY INTEREST COMPANY",
-	"COMMUNITY INTEREST P.L.C",
-	"COMMUNITY INTEREST PLC",
-	"COMMUNITY INTEREST PUBLIC LIMITED COMPANY",
-	"CWMNI BUDDIANT C.C.C",
-	"CWMNI BUDDIANT CCC",
-	"CWMNI BUDDIANT CYMUNEDOL C.C.C",
-	"CWMNI BUDDIANT CYMUNEDOL CCC",
-	"CWMNI BUDDIANT CYMUNEDOL CYHOEDDUS CYFYNGEDIG",
-	"CWMNI BUDDIANT CYMUNEDOL",
-	"CWMNI BUDDSODDIA CHYFALAF NEWIDIOL",
-	"CWMNI BUDDSODDIANT PENAGORED",
-	"CWMNI CELL GWARCHODEDIG",
-	"CWMNI CYFYNGEDIG CYHOEDDUS",
-	"CYF",
-	"CYFYNGEDIG",
-	"EEIG",
-	"EESV",
-	"EOFG",
-	"EOOS",
-	"EUROPEAN ECONOMIC INTEREST GROUPING",
-	"GEIE",
-	"GELE",
-	"ICVC",
-	"INVESTMENT COMPANY WITH VARIABLE CAPITAL",
-	"L.P",
-	"L.T.D",
-	"LIMITED - THE",
-	"LIMITED LIABILITY PARTNERSHIP",
-	"LIMITED PARTNERSHIP",
-	"LIMITED THE",
-	"LIMITED",
-	"LIMITED-THE",
-	"LIMITED...THE",
-	"LIMITED..THE",
-	"LIMITED.THE",
-	"LLP",
-	"LP",
-	"LTD",
-	"LTD...THE",
-	"LTD..THE",
-	"LTD.THE",
-	"OEIC",
-	"OPEN-ENDED INVESTMENT COMPANY",
-	"P.L.C",
-	"PAC",
-	"PARTNERIAETH ATEBOLRWYDD CYFYNGEDIG",
-	"PARTNERIAETH CYFYNGEDIG",
-	"PCC LIMITED",
-	"PCC LTD",
-	"PCC",
-	"PLC",
-	"PROTECTED CELL COMPANY",
-	"PUBLIC LIMITED COMPANY .THE",
-	"PUBLIC LIMITED COMPANY THE",
-	"PUBLIC LIMITED COMPANY",
-	"PUBLIC LIMITED COMPANY.THE",
-	"UNLIMITED",
-	"UNLTD",
-}
 
 var (
 	filename1 = "company-errors/error-posting-request.txt"
@@ -358,30 +279,6 @@ func (c *connections) sendToES(companies *[]*mongoCompany, length int) {
 
 // ---------------------------------------------------------------------------
 
-//Split company name into nameStart and nameEnding in order to remove common name endings
-func splitCompanyNameEndings(name string) (string, string) {
-	var nameStart, nameEnding string
-
-	nameStart = name
-
-	//Strip trailing non-word characters [^a-zA-Z0-9_]
-	stripped := nonWordEndRegex.ReplaceAllString(name, "")
-
-	//Scan company name for name ending and remove suffix
-	for _, cne := range companyNameEndings {
-		if strings.HasSuffix(stripped, cne) {
-			nameStart = strings.TrimSuffix(stripped, " "+cne)
-			// Keep the actual name ending by extracted the name start
-			nameEnding = name[len(nameStart):]
-			break
-		}
-	}
-
-	return nameStart, nameEnding
-}
-
-// ---------------------------------------------------------------------------
-
 /*
 Pass in a reference to mongoCompany, as golang is pass-by-value. This version, golang
 will create a copy of mongoCompany on the stack for every call (which is good, as it
@@ -408,7 +305,9 @@ func (c *connections) mapResult(source *mongoCompany) *esCompany {
 	}
 
 	name := source.Data.CompanyName
-	nameStart, nameEnding := splitCompanyNameEndings(source.Data.CompanyName)
+
+	f := &format.Format{}
+	nameStart, nameEnding := f.SplitCompanyNameEndings(source.Data.CompanyName)
 
 	items := esItem{
 		CompanyStatus:       source.Data.CompanyStatus,
